requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
requests==2.32.3
PyYAML==6.0.2
